Fixed issue where external links were not open using InAppBrowser.
Once upon a time, Cordova used to open external links in the same
Cordova WebView that was running the application. When a link was open
and the user clicked the “back” button, the last displayed page was
shown exactly as it was before the user left it. 
In the newer version, this has changed. Nowadays, the external links are
opened, by default, using the Cordova WebView if the URL is in your
app’s whitelist. URLs that aren’t on your whitelist are opened using the
InAppBrowser API. If you don’t manage the links in the right way, or if
the user taps a link that is shown in the InAppBrowser or the system and
then chooses to go back, all the jQuery Mobile enhancements are lost.
This behavior happens because the CSS and JavaScript files are loaded by
the main page, and the following ones are loaded using AJAX.